Fix GCC/Linux compilation and runtime issues for cross-platform support
- Fix MSVC-specific constructs for GCC compatibility: std::exception(string)
  to std::runtime_error, byte to unsigned char, MAXINT to INT_MAX,
  va_list{} initialization, rvalue-to-lvalue binding, typename for
  dependent types, std::ifstream(wstring) via std::filesystem::path
- Add bool to std::variant type list in itemdata.h for GCC
- Add VLC 3.x/4.x API compatibility layer (SDL_PLATFORM_WIN32 guards)
- Resolve spdlog/fmt ABI mismatch via FetchContent (spdlog v1.10.0)
- Use -idirafter for vendored Libs/ on non-Windows so FetchContent
  headers take priority
- Add platform guards for _ftelli64/_fseeki64 vs ftello64/fseeko64
  and _lock_file/_unlock_file vs flockfile/funlockfile
- Add platform guard for MultiByteToWideChar vs std::wstring_convert
- Force -O2 on stb_impl.cpp for GCC (SSE intrinsics require optimization)
- Fix static const auto to static constexpr int for ODR-use in Starfield
- Fix missing default keyboard/mouse bindings in InitDefaultConfiguration
- Add HiDPI-aware blit in OpenGL Present() using SDL_GetWindowSizeInPixels
- Add SDL_HINT_VIDEO_WAYLAND_SCALE_TO_DISPLAY hint before SDL_Init
- Fix liblz4_static_d.lib reference in vcxproj (lib was removed)

+++ TombEngine/Game/effects/Streamer.cpp
@@ -108,7 +108,7 @@ namespace TEN::Effects::Streamer
 		auto& segment = GetSegment();
 
 		// Avoid creating "clipped" streamers by clamping max life according to max segment count.
-		int lifeMax = (int)std::min(round(life * FPS), (float)SEGMENT_COUNT_MAX);
+		int lifeMax = (int)std::min((float)round(life * FPS), (float)SEGMENT_COUNT_MAX);
 
 		float alpha = (float(segmentCount * SEGMENT_SPAWN_INTERVAL_TIME) / (float)lifeMax) * FADE_IN_COEFF;
 		float opacityMax = EaseInOutSine(colorEnd.w, colorStart.w, alpha);
